requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
langchain-chroma
psycopg2-binary
pandas
pydantic
pydantic-settings
python-multipart
tavily-python
python-dotenv